organize hardware drivers
Move ADS1115, RTC, RF, USB-PD, and display code into the shallow drivers boundary. Retain compatibility headers at the repository root and compile Arduino-only implementations out of Linux targets.
